Update processor.coffee
This is a naive proposal for fixing a bug I found in my research documented at https://github.com/postcss/postcss-import/issues/19#issuecomment-71777987. Also it may be that the bug is with `postcss-import`. But technically this will make the problem go away, I just don't know why it happened in the first place. Very likely @ai can explain the issue and offer a better solution.