DynJS compatibility
DynJS (dynjs.org) is a JVM-based ECMAscript engine, which currently fails to run this module as it expects array sort functions to always return integer values, not floats. This minor change will still work with other more lenient engines but allows autoprefixer to run in DynJS, which in turn allows this module to work in Nodyn (http://nodyn.io/).