Add hack for unitless 0 basis in IE10/IE11 shorthand flex (#749)
Only applies when there are three components in the flex property.

This is because when there's one or two components only, it's
ambiguous whether whether the first component being a unitless 0 is
intended to be flex-grow or flex-basis. Browsers resolve this by
treating such cases as a flex-grow of 0.

Note with IE11 that even though it understands the unprefixed flex
property, we don't apply the hack there. Instead we let it treat the
unprefixed version as invalid, and fall back onto -ms-flex.

Finally, although some CSS minifiers may revert the 0px to 0, we
don't use 0% instead, even though many of these won't touch 0%.
See https://bugs.chromium.org/p/chromium/issues/detail?id=495306
for why this is the case.

Workaround for:
https://github.com/philipwalton/flexbugs#4-flex-shorthand-declarations-with-unitless-flex-basis-values-are-ignored

References:
https://www.w3.org/TR/2012/WD-css3-flexbox-20120322/#flexibility
https://www.w3.org/TR/css-flexbox-1/#flex-property
https://www.w3.org/TR/2016/CR-css-values-3-20160929/#component-combinators