Update stylesheet test cases to fit the new duplicate rules feature
As we have duplicate grid-area rules support, the order of rules is changed now. Before that the pack of duplicate rules with prefixes was inserted after the last grid-area rule, but now each duplicate rule is inserted right after the matched rule.

We also preserve the order of prefixed grid-area rules and merge the rules with the same media parameters